update file
function that searches

for a value in an array

of integers using

the Linear search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,15 +10,15 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array == NULL)
 	{
 		return (-1);
 	}
-	for (i = 0; i < size ; i++)
+	for (i = 0; i < (int)size ; i++)
 	{
-		printf("Value checked array [%ld] = [%d]\n", i, (int)array[i]);
+		printf("Value checked array [%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
